pyproject.toml
[project]
name = "manga-ocr"
authors = [
    {name = "Maciej Budyś", email = "kha-white@mail.com"},
]
description = "OCR for Japanese manga"
readme = "README.md"
requires-python = ">=3.11"
license = {file = "LICENSE"}
classifiers = [
    "Programming Language :: Python :: 3",
]

dynamic = ["version"]
dependencies = [
    "albumentationsx",
    "budoux",
    "fire",
    "fonttools",
    "fugashi",
    "html2image",
    "jaconv",
    "loguru",
    "numpy",
    "opencv-python",
    "pandas",
    "Pillow",
    "pyperclip",
    "torch",
    "transformers",
    "unidic_lite",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "ruff",
    "pydantic-settings",
    "typer",
    "wandb",
    "pictex",
]

[project.urls]
Homepage = "https://github.com/kha-white/manga-ocr"

[build-system]
requires = ["setuptools >= 61.0.0", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["manga_ocr"]

[tool.setuptools.dynamic]
version = {attr = "manga_ocr._version.__version__"}

[tool.ruff]
line-length = 120
indent-width = 4

[project.scripts]
manga_ocr = "manga_ocr.__main__:main"

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true


[tool.uv.sources]
torch = [  
{ index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [  
{ index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
